% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{allowed_amino_acids}
\alias{allowed_amino_acids}
\title{String of allowed amino acids}
\usage{
allowed_amino_acids()
}
\description{
Returns a data.frame of allowed amino acid single-letter codes. These come
from IUPAC (International Union of Pure and Applied Chemistry),
\href{https://www.bioinformatics.org/sms/iupac.html}{see here} for details.
Variants defined in the \code{variant_string} column of the counts data.frame
must be one of these single-letter codes, or otherwise the underscore (_),
pipe (|) or forward slash (/) symbols. Underscores will be removed but are
allowed for improved readability. The pipe and forward slash symbols
represent phased or unphased mixed calls, respectively.
}
