% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{allowed_country_names}
\alias{allowed_country_names}
\title{Vector of allowed country names}
\usage{
allowed_country_names()
}
\description{
Get a vector of allowed country names for use in the survey table. These
names were taken from GADM 4.1.
}
