% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_input}
\alias{example_input}
\title{Example Input Dataset}
\format{
A list containing three data.frames
}
\usage{
example_input
}
\description{
A list containing the three data.frames that make up a STAVE object. These
are correctly formatted, so can be used as a reference to when formatting
data. The studies themselves come from a much reduced version of the WWARN
data on antimalaria resistance.
}
\examples{
data("example_input", package = "STAVE")
}
\keyword{datasets}
