% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{STAVE_object}
\alias{STAVE_object}
\title{STAVE data object (R6 class)}
\description{
TODO
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-STAVE_object-print}{\code{STAVE_object$print()}}
\item \href{#method-STAVE_object-append_data}{\code{STAVE_object$append_data()}}
\item \href{#method-STAVE_object-get_prevalence}{\code{STAVE_object$get_prevalence()}}
\item \href{#method-STAVE_object-get_variants}{\code{STAVE_object$get_variants()}}
\item \href{#method-STAVE_object-export_xlsx}{\code{STAVE_object$export_xlsx()}}
\item \href{#method-STAVE_object-clone}{\code{STAVE_object$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-STAVE_object-print"></a>}}
\if{latex}{\out{\hypertarget{method-STAVE_object-print}{}}}
\subsection{Method \code{print()}}{
Custom print method to control console output
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{STAVE_object$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-STAVE_object-append_data"></a>}}
\if{latex}{\out{\hypertarget{method-STAVE_object-append_data}{}}}
\subsection{Method \code{append_data()}}{
Append new data
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{STAVE_object$append_data(
  studies_dataframe,
  surveys_dataframe,
  counts_dataframe
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{studies_dataframe}}{TODO}

\item{\code{surveys_dataframe}}{TODO}

\item{\code{counts_dataframe}}{TODO}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-STAVE_object-get_prevalence"></a>}}
\if{latex}{\out{\hypertarget{method-STAVE_object-get_prevalence}{}}}
\subsection{Method \code{get_prevalence()}}{
Calculate prevalence
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{STAVE_object$get_prevalence(variant)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{variant}}{TODO}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-STAVE_object-get_variants"></a>}}
\if{latex}{\out{\hypertarget{method-STAVE_object-get_variants}{}}}
\subsection{Method \code{get_variants()}}{
Return a table of all variants present in the data object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{STAVE_object$get_variants(report_haplo = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{report_haplo}}{(Boolean) if TRUE then list all haplotypes.
Otherwise, list in locus-by-locus format. Defaults to FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-STAVE_object-export_xlsx"></a>}}
\if{latex}{\out{\hypertarget{method-STAVE_object-export_xlsx}{}}}
\subsection{Method \code{export_xlsx()}}{
Export full data object as an Excel spreadsheet (.xlsx format).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{STAVE_object$export_xlsx(file_path)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{file_path}}{the full path and name of the output file.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-STAVE_object-clone"></a>}}
\if{latex}{\out{\hypertarget{method-STAVE_object-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{STAVE_object$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
