% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{STAVE_object}
\alias{STAVE_object}
\title{STAVE data object (R6 class)}
\description{
The main class that stores the data and is responsible for all data input,
output, and processing functions. Most of the functionality of the STAVE
package is through this class in the form of member functions.
}
\details{
The raw data are stored as private variables within this object, meaning they
cannot (or should not) be edited directly. Rather, tables can be extracted
using \code{get_counts()} and similarly for other tables. The three tables
are:
  \enumerate{
    \item studies: Information on where the data came from, for example a url
    and author names. Each study is indexed with a unique study_ID.
    \item surveys: Information on the surveys represented within a study. A
    survey is defined here as a discrete instance of data collection, which
    includes information on geography (latitude and longitude) and collection
    time. Surveys are given survey_IDs and are linked to a particular study
    through the study_ID.
    \item counts: The actual genetic information, which is linked to a
    particular survey through the survey_ID. Genetic variants are encoded in
    character strings that must follow a specified format, and the number of
    times this variant was observed among the total sample is stored in
    columns.
  }
This combination of linked tables allows efficient and flexible encoding of
variants, while avoiding unnecessary duplication of information.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-STAVE_object-print}{\code{STAVE_object$print()}}
\item \href{#method-STAVE_object-get_studies}{\code{STAVE_object$get_studies()}}
\item \href{#method-STAVE_object-get_surveys}{\code{STAVE_object$get_surveys()}}
\item \href{#method-STAVE_object-get_counts}{\code{STAVE_object$get_counts()}}
\item \href{#method-STAVE_object-get_version}{\code{STAVE_object$get_version()}}
\item \href{#method-STAVE_object-append_data}{\code{STAVE_object$append_data()}}
\item \href{#method-STAVE_object-get_prevalence}{\code{STAVE_object$get_prevalence()}}
\item \href{#method-STAVE_object-get_variants}{\code{STAVE_object$get_variants()}}
\item \href{#method-STAVE_object-drop_study}{\code{STAVE_object$drop_study()}}
\item \href{#method-STAVE_object-clone}{\code{STAVE_object$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-STAVE_object-print"></a>}}
\if{latex}{\out{\hypertarget{method-STAVE_object-print}{}}}
\subsection{Method \code{print()}}{
Custom print method to control console output
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{STAVE_object$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-STAVE_object-get_studies"></a>}}
\if{latex}{\out{\hypertarget{method-STAVE_object-get_studies}{}}}
\subsection{Method \code{get_studies()}}{
Extract the studies data.frames stored within the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{STAVE_object$get_studies()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-STAVE_object-get_surveys"></a>}}
\if{latex}{\out{\hypertarget{method-STAVE_object-get_surveys}{}}}
\subsection{Method \code{get_surveys()}}{
Extract the surveys data.frames stored within the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{STAVE_object$get_surveys()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-STAVE_object-get_counts"></a>}}
\if{latex}{\out{\hypertarget{method-STAVE_object-get_counts}{}}}
\subsection{Method \code{get_counts()}}{
Extract the counts data.frames stored within the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{STAVE_object$get_counts()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-STAVE_object-get_version"></a>}}
\if{latex}{\out{\hypertarget{method-STAVE_object-get_version}{}}}
\subsection{Method \code{get_version()}}{
Extract the version number of the STAVE object. This is important as
member functions of a STAVE object are directly linked to the object
itself, and will not be updated by updating the version of the package in
your environment. To update a STAVE object to a new package version, you
should first extract the data and then load into a new STAVE object
created with the most recent version.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{STAVE_object$get_version()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-STAVE_object-append_data"></a>}}
\if{latex}{\out{\hypertarget{method-STAVE_object-append_data}{}}}
\subsection{Method \code{append_data()}}{
Append new data
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{STAVE_object$append_data(
  studies_dataframe,
  surveys_dataframe,
  counts_dataframe
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{studies_dataframe}}{a data.frame containing information at the study
level. This data.frame must have the following columns: study_ID,
study_name, study_type, authors, publication_year, url}

\item{\code{surveys_dataframe}}{a data.frame containing information at the
survey level. This data.frame must have the following columns:
study_key, survey_ID, country_name, site_name, lat, lon, spatial_notes,
collection_start, collection_end, time_notes. The study_key element
must correspond to a study_ID in the studies_dataframe.}

\item{\code{counts_dataframe}}{a data.frame of genetic information. Must contain
the following columns: survey_key, variant_string, variant_num,
total_num. The survey_key element must correspond to a valid survey_ID
in the surveys_dataframe.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-STAVE_object-get_prevalence"></a>}}
\if{latex}{\out{\hypertarget{method-STAVE_object-get_prevalence}{}}}
\subsection{Method \code{get_prevalence()}}{
Calculate prevalence
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{STAVE_object$get_prevalence(
  target_variant,
  keep_ambiguous = FALSE,
  prev_from_min = TRUE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{target_variant}}{the name of the variant on which we want to
calculate prevalence, for example crt:72:C. Note that there can be no
heterozygous calls within this name.}

\item{\code{keep_ambiguous}}{there may be variants in the data for which the
target_variant could be in the sample, but this cannot be proven
conclusively. For example, the sequence A_A_A may be a match to the
sequence A/C_A/C_A or it may not, these are unphased genotypes so we
cannot be sure. If \code{keep_ambiguous = TRUE} then both a min and a
max numerator are reported that either exclude all ambiguous calls
(min) or include all ambiguous calls (max). If \code{FALSE} (the
default) then only the min is reported.}

\item{\code{prev_from_min}}{the output object includes a point estimate of the
prevalence along with exact binomial confidence intervals. These must
be calculated from one of \code{numerator_min} or \code{numerator_max}
in the case of ambiguous calls. This argument sets which one of these
numerators is used in the calculation.

@import dplyr}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-STAVE_object-get_variants"></a>}}
\if{latex}{\out{\hypertarget{method-STAVE_object-get_variants}{}}}
\subsection{Method \code{get_variants()}}{
Return a vector of all variants present in the data object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{STAVE_object$get_variants(report_haplo = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{report_haplo}}{(Boolean) if TRUE then list all haplotypes.
Otherwise, list in locus-by-locus format. Defaults to FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-STAVE_object-drop_study"></a>}}
\if{latex}{\out{\hypertarget{method-STAVE_object-drop_study}{}}}
\subsection{Method \code{drop_study()}}{
Drop one or more study_IDs from the data. This will drop from all
internally stored data objects, including the corresponding surveys and
counts data.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{STAVE_object$drop_study(drop_study_ID)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{drop_study_ID}}{a vector of study_IDs to drop from all data objects.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-STAVE_object-clone"></a>}}
\if{latex}{\out{\hypertarget{method-STAVE_object-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{STAVE_object$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
